Added update filters and create filters functionality
- Added functionality to send updated filters and created filters to
the Tunestumbler API based on the state `new` flag. `new: true` means
that the filter is newly being added, and `new: false` means that the
filter is modified/unmodified
- Note: `filtersId` for new filters is `null`, since this is
generated sever side
- Note: `multireddit`, `priority`, `allowNSFWFlag` are set to default
values since controls are not yet created for them
- Added `shouldUpdateFilters` and `shouldCreateFilters` to component
states to keep track of whether to expect filters to be updated or
created, which are used to determine if the user should be redirected
to the `/results/new` route
- Updated `/results` routes
- `/filters` route will only redirect to `/results` route only if
there is at least one filters control. This filters control may
either be a newly created filter, an updated filter, or an unmodified
previously saved filter. Unmodified filters are just sent to the
`/updated` API endpoint for easier implementation.
- Added the `valid` controls property to the `filters` controls, so
that the UI for the controls turns red when any of the inputs are
incorrect
- Added uuid to the `Auxiliary` component of each filter controls to
fix one React warning of not having a unique key
- Fixed filters axios `PUT` body

temp