Updated Login container, Input and Button components
- Changed Login page css styling
- Implemented React-Bootstrap’s `Form` and `Button`, since it looks
better and has the same functionality
- Added Login form labels in place of placeholders, since it is more
intuitive
- There is a bug with React-Bootstrap’s `Form`, where it does not
properly handle the `onSubmit` event. So as a workaround, a separate
`onEnter` function is implemented to handle when the user hits the
`Enter` key on the Login form. (This triggers the `onLogin` redux
function, which propagates the email and password values to the Redux
store and action creators)
- Keys have been added to the `Email` and `Password` inputs
- Disabled default html form validation for email and password
- Removed `shouldValidate` property since this is redundant
- Removed `elementDisplay` property from Input component since won’t
need to apply sizes to React-Bootstrap controls
- Removed `textarea` portion from Input component since it is
redundant
- Changed `input` element to `Form.Control` element in the Input
component since all inputs will be part of a form anyways
- Updated Button component to use React-Bootstrap’s Button component
since it looks nicer
- Modified Button component `className` property so that container
specific props can be passed to it