Updated Filters container, Input component, and Navigation styling
- Added error message that will be used to alert user when a form
input is invalid
- Modified filter forms to use React-Bootstraps `<Form>`
- Added a large breakpoint on the forms
- Removed unit tests for `ControlsForm` and `FiltersForm` for now
because these components rely heavily on the `FiltersContainer`
state. Separating these forms into their own components mean that a
large portion of the state would have to be passed to both of these
components and also the state would need to be passed up the
component hierarchy, back to the container where the all values of
the container state would need to be deep copied. Deep copying would
happen every time something in the form is updated, which I think
will take up  a lot of browser resources. Because of the burden this
 would put on the client side, I think it’s best to not separate out
functionality until I can think of a better approach
- Fixed the `NavigationBar` background color not filling the row when
the user is on a small viewport and needs to scroll to the overflowed
`NavigationItems`
- Changed the `select` input case of the `Input` component to use
React-Bootstraps select `<Form.Control>`